chore: move getImageDescription to a standalone file
