+++ gallery/src/lib/browser.ts
@@ -1,0 +1,1 @@
+export { getImageDescription } from '../utils/descriptions';
